scripts/requirements.txt
requests>=2.31.0
orjson>=3.9.10
polygon-api-client>=1.12.3
yfinance>=0.2.36
curl-cffi>=0.5.11
pandas>=2.2.2
exchange-calendars>=4.5
dateparser>=1.2.0
cachetools>=5.3.3
numpy>=1.26.4
numba>=0.59.1
scipy>=1.13.0 